Homoscedasticity tests bug fix
Homoscedasticity was inversely evaluated. Both Bartlett's and Filgner's tests test the null hypothesis that the variances in each of the groups are the same. This means that when p-value < threshold (usually 0.05) the null hyphothesis is rejected and, therefore, the variance is considered different in each group.

Incorrectly, the condition used to consider the variance different was p-value >= threshold.